+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Replace binary values in sql value generation with `<binary>` identifier.
 - Fix a couple sql unicode bugs.
 - Use single quotes for sql string literals, and do not quote integers.
+- Fixed a bug in the AsyncMultiQuery that would prevent returning all results in the case when an `OR` query was used with an offset and some entities matched more than one branch of the `OR` query.
 
 ## v0.9.11